+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 * `auto disable s r v lights` setting: Automatically turn SRV lights off when
   deploying one. Default: true. (#133)
 
+### Fixed
+
+* `auto enter station services` option. (#142)
+
 ## SpashAttack 7.2.2
 
 * Fixed getting current jump range from EDDI; no longer fails on the first try,